Move set_scan_status() from attack.c to reporting.c
Do not store the message in redis main kb.
